# Conflicts:
#	src/sg/edu/nus/comp/cs4218/impl/app/MvApplication.java

+++ src/sg/edu/nus/comp/cs4218/impl/app/MvApplication.java
@@ -1,23 +1,29 @@
 package sg.edu.nus.comp.cs4218.impl.app;
 
-import sg.edu.nus.comp.cs4218.app.MvInterface;
-import sg.edu.nus.comp.cs4218.exception.InvalidArgsException;
-import sg.edu.nus.comp.cs4218.exception.MvException;
-import sg.edu.nus.comp.cs4218.impl.parser.MvArgsParser;
-import sg.edu.nus.comp.cs4218.impl.util.IOUtils;
+import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.ERR_CANNOT_OVERWRITE;
+import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.ERR_FILE_NOT_FOUND;
+import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.ERR_NO_FILE_ARGS;
+import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.ERR_NO_OSTREAM;
+import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.ERR_NULL_ARGS;
 
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.DirectoryNotEmptyException;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.List;
 
-import static sg.edu.nus.comp.cs4218.impl.util.ErrorConstants.*;
+import sg.edu.nus.comp.cs4218.app.MvInterface;
+import sg.edu.nus.comp.cs4218.exception.InvalidArgsException;
+import sg.edu.nus.comp.cs4218.exception.MvException;
+import sg.edu.nus.comp.cs4218.impl.parser.MvArgsParser;
+import sg.edu.nus.comp.cs4218.impl.util.IOUtils;
 
+@SuppressWarnings("PMD.PreserveStackTrace")
 public class MvApplication implements MvInterface {
 
     private Boolean shouldOverwrite = false;
@@ -28,7 +34,7 @@ public class MvApplication implements MvInterface {
             Path source = IOUtils.resolveFilePath(srcFile);
             Files.move(source, source.resolveSibling(destFile));
         } catch (Exception e) {
-            throw (MvException) new MvException(ERR_FILE_NOT_FOUND).initCause(e);
+            throw new Exception(ERR_FILE_NOT_FOUND);
         }
 
         return null;
@@ -36,17 +42,22 @@ public class MvApplication implements MvInterface {
 
     @Override
     public String mvFilesToFolder(String destFolder, String... fileName) throws Exception {
+
         for (String srcPath: fileName){
             if (shouldOverwrite) {
                 // Can avoid this with assumption that target operand is always a directory
-                if (!Files.isDirectory(IOUtils.resolveFilePath(destFolder))) {
+                if (!Files.isDirectory(IOUtils.resolveFilePath(destFolder)) && fileName.length == 1) {
                     FileOutputStream outputStream = new FileOutputStream(IOUtils.resolveFilePath(destFolder).toFile());//NOPMD
                     byte[] strToBytes = Files.readAllBytes(IOUtils.resolveFilePath(srcPath));
                     outputStream.write(strToBytes);
                     outputStream.close();
 
-                    Files.delete(IOUtils.resolveFilePath(srcPath));
+                    if (!IOUtils.resolveFilePath(srcPath).equals(IOUtils.resolveFilePath(destFolder))) {
+                        Files.delete(IOUtils.resolveFilePath(srcPath));
+                    }
                     return null;
+                } else if (!Files.isDirectory(IOUtils.resolveFilePath(destFolder)) && fileName.length > 1) {
+                    throw new Exception("'" + srcPath + "' is not a directory.");
                 }
                 try {
                     // Assumption: Replacement doesn't work when a directory is being moved and target directory is non-empty and same name
@@ -55,16 +66,21 @@ public class MvApplication implements MvInterface {
                                     IOUtils.resolveFilePath(srcPath).getFileName().toString()),
                             StandardCopyOption.REPLACE_EXISTING);
                 } catch (DirectoryNotEmptyException dnee) {
-                    throw (MvException) new MvException(ERR_CANNOT_OVERWRITE).initCause(dnee);
+                    throw new Exception(ERR_CANNOT_OVERWRITE + " non-empty directory: " + destFolder);
                 }
 
             } else {
                 // if overwriting is not allowed then only allow possibility of moving if its directory
                 if (Files.isDirectory(IOUtils.resolveFilePath(destFolder))) {
-                    Files.move(IOUtils.resolveFilePath(srcPath),
-                            Paths.get(IOUtils.resolveFilePath(destFolder).toString(), srcPath));
+                    try {
+                        Files.move(IOUtils.resolveFilePath(srcPath),
+                                Paths.get(IOUtils.resolveFilePath(destFolder).toString(), srcPath));
+                    } catch (FileAlreadyExistsException faee) {
+                        throw new Exception(ERR_CANNOT_OVERWRITE + " file already exists: " + srcPath); //NOPMD
+                    }
+
                 } else {
-                    throw new MvException(ERR_NOT_MOVABLE);
+                    throw new Exception("'" + srcPath + "' is a file and replacement not allowed.");
                 }
             }
         }
@@ -77,6 +93,10 @@ public class MvApplication implements MvInterface {
             throw new MvException(ERR_NULL_ARGS);
         }
 
+        if (stdout == null) {
+            throw new MvException(ERR_NO_OSTREAM);
+        }
+
         // Assumption: Will assume multiple args passed when regex is used with only first arg passed
         if (args.length < 2) {
             throw new MvException(ERR_NO_FILE_ARGS);
@@ -86,7 +106,7 @@ public class MvApplication implements MvInterface {
         try {
             parser.parse(args);
         } catch (InvalidArgsException e) {
-            throw (MvException) new MvException(e.getMessage()).initCause(e);
+            throw new MvException(e.getMessage());
         }
 
         shouldOverwrite = parser.shouldOverwrite();
@@ -100,7 +120,7 @@ public class MvApplication implements MvInterface {
 
         for (String sourceOperand : sourceOperands) {
             if (!Files.exists(IOUtils.resolveFilePath(sourceOperand))) {
-                throw new MvException(ERR_FILE_NOT_FOUND + ": " + sourceOperand);
+                throw new MvException("Cannot find '" + sourceOperand + "'. " +  ERR_FILE_NOT_FOUND + ".");
             }
         }
 
@@ -108,10 +128,13 @@ public class MvApplication implements MvInterface {
             if (Files.exists(IOUtils.resolveFilePath(targetOperand))) {
                 mvFilesToFolder(targetOperand, sourceOperands.toArray(new String[0]));
             } else {
+                if (sourceOperands.size() > 1) {
+                    throw new Exception("can't rename multiple files");
+                }
                 mvSrcFileToDestFile(sourceOperands.get(0), targetOperand);
             }
         }  catch (Exception e) {
-            throw (MvException) new MvException(e.getMessage()).initCause(e);
+            throw new MvException(e.getMessage());
         }
 
     }
